progress: 32.30% of code matched (4742/8216 functions, 4975/6137 objects)
Source-commit: e1742e0b4ecded5a7f96d927e50afcf4b17f5249

+++ src/game/game_fn_800A2414.c
@@ -16,7 +16,7 @@ typedef struct Object800A2414 {
 
 void fn_800A2414(Object800A2414* object)
 {
-    if (object->state == 1) {
+    if ((int)object->state == 1) {
         object->count--;
     }
 }
