+++ src/game/game_fn_8018168C.c
@@ -14,7 +14,7 @@ extern s16 lbl_80606360[];
 
 void fn_8018168C(Entry* dst, const Coord* src, s16 angle, int scale)
 {
-    s16 index = angle;
+    int index = angle;
     dst->position.x = src->x + ((lbl_80606360[index] * scale) >> 7);
     dst->position.y = src->y + ((lbl_80606360[index + 0x40] * scale) >> 7);
 }
